requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
celery
redis==4.5.1
django-cors-headers==3.13.0
python-dotenv==0.21.1
django-celery-beat==2.5.0
requests==2.31.0

# Development and Testing
pytest==7.4.0
pytest-django==4.5.2
pytest-cov==4.1.0
pytest-celery==0.0.0
factory-boy==3.3.0
faker==19.3.0

# Code Quality
flake8==6.0.0
black==23.7.0
isort==5.12.0
pre-commit==3.3.3
